Render parsed Org document
Observe that the container where the markup is rendered is drawn before the source code arrives.

This is horrible, but if the container is rendered on the same loop that triggers the port call the JS code might execute before the call to view, causing an error when the container is not found.